+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] / 2023-04-08
+### Features
+- RibbonItem - CreateCopy
+- RibbonPanel - MoveToRibbonTab
+### Utils
+- RibbonTabUtils - CreatePanel, RemovePanel
+- RibbonModifyUtils - CreatePanel, RemovePanel
+
 ## [0.4.0] / 2022-08-23
 ### Features
 - Bitmap to BitmapSource perfect pixel